embedder/requirements.txt
qdrant-client
voyageai
fastembed
python-dotenv
boto3
tqdm
